pyproject.toml
[tool.poetry]
name = "oc-cg-p10"
version = "0.1.0"
description = ""
authors = ["Ciran GÜRBÜZ <ciran.gurbuz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
requests = "^2.31.0"
djangorestframework = "^3.14.0"
black = "^23.7.0"
flake8 = "^6.1.0"
djangorestframework-simplejwt = "^5.3.0"
drf-nested-routers = "^0.93.4"
docformatter = "^1.7.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79